feat(billets-web): :art: implemented rotate spinner
