+++ packages/ocean-road/src/spinner/spinner.styled.ts
@@ -1,6 +1,9 @@
 import styled from '@emotion/styled'
+import { motion } from 'framer-motion'
 import { LoaderCircle } from 'lucide-react'
 import { semantics } from '../tokens'
+
+const MotionIcon = motion.create(LoaderCircle)
 
 export const StyledPageLoadingWrapper = styled.div`
   position: absolute;
@@ -13,7 +16,7 @@ export const StyledPageLoadingWrapper = styled.div`
   justify-content: center;
 `
 
-export const StyledSpinner = styled(LoaderCircle)`
+export const StyledSpinner = styled(MotionIcon)`
   color: ${semantics.color.foreground[1]};
   width: 32px;
   height: 32px;
